Bug #2043970: MAAS 3.2.9 creates for Calico Interfaces 80.000 fabrics
Redo the fix for bug #2043970.

In some cases we need to create VLANs/fabrics for interfaces that don't have
links. For example for VLAN interfaces, often it's the VLAN interface that has
an address, while the underlying physical interface often is without any
address. In that case, the physical interface should have a VLAN that is on the
same fabric as the VLAN interface.

Also removed the check whether a machine is a controller. A controller is not
special. It's a deployed machine. If Calico is installed on a controller, we
should not create useless fabrics.

(cherry picked from commit 94dac1082c3a3958e7c01f29e1d0320a775633ae)
(cherry picked from commit 3c34344be9f9ab6f283a505193c678ff039c17ef)